requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
streamlit>=1.28.0
pydantic>=2.0.0